# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: phpGRC
  version: 0.4.7
  description: |
    Notes:
    - The Metrics Dashboard is an internal, admin-only endpoint (`GET /dashboard/kpis`, alias `GET /metrics/dashboard`) and is not part of this public OpenAPI. Both return identical shapes.
  license:
    identifier: MIT
    name: MIT
  x-logo:
    url: "https://phpgrc.gruntlabs.net/api/images/phpGRC-light-horizontal-trans.png"
    altText: "phpGRC"
    href: "https://phpgrc.gruntlabs.net/api/docs"
servers:
  - url: /
...
security:
  - sanctum: []
paths:
...
  /exports:
    post:
      operationId: exportsCreateLegacy
      tags: [Exports]
      summary: Create export (legacy)
      description: Creates an export job using the legacy request format.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportCreateLegacyRequest'
      responses:
        '200':
          description: Created/Queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportCreateResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
  /exports/{type}:
    post:
      operationId: exportsCreateByType
      tags: [Exports]
      summary: Create export by type
      description: Creates an export job for the given type, with optional parameters.
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
            enum: [csv, json, pdf]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportCreateTypeRequest'
      responses:
        '200':
          description: Created/Queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportCreateResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    sanctum:
      type: http
      scheme: bearer
      bearerFormat: Token
...
  responses:
    Unauthenticated:
      description: Unauthenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Unauthenticated'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ForbiddenError'
    ValidationFailed:
      description: Validation errors
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationFailed'
...
    RateLimited:
      description: Rate limited
      headers:
        Retry-After:
          description: Seconds until a new request is allowed.
          schema:
            type: integer
            minimum: 1
        X-RateLimit-Limit:
          description: Requests allowed in the window.
          schema:
            type: integer
            minimum: 1
        X-RateLimit-Remaining:
          description: Requests remaining in the current window.
          schema:
            type: integer
            minimum: 0
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RateLimited'
  schemas:
...
    ErrorSimple:
      type: object
      required: [ok, code]
      properties:
        ok:
          type: boolean
          const: false
        code:
          type: string
        message:
          oneOf:
            - type: string
            - type: 'null'
        note:
          oneOf:
            - type: string
            - type: 'null'
    Unauthenticated:
      allOf:
        - $ref: '#/components/schemas/ErrorSimple'
        - type: object
          properties:
            code:
              type: string
              enum: [UNAUTHENTICATED]
    ForbiddenError:
      allOf:
        - $ref: '#/components/schemas/ErrorSimple'
        - type: object
          properties:
            code:
              type: string
              enum: [UNAUTHORIZED]
    ValidationFailed:
      type: object
      required: [ok, code, message, errors]
      properties:
        ok:
          type: boolean
          const: false
        code:
          type: string
          enum: [VALIDATION_FAILED]
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
...
    RateLimited:
      type: object
      required: [ok, code, retry_after]
      properties:
        ok:
          type: boolean
          const: false
        code:
          type: string
          enum: [RATE_LIMITED]
        retry_after:
          type: integer
          minimum: 1
...
    ExportCreateLegacyRequest:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [csv, json, pdf]
        params:
          type: object
          additionalProperties: true
    ExportCreateTypeRequest:
      type: object
      properties:
        params:
          type: object
          additionalProperties: true
    ExportCreateResponse:
      type: object
      required: [ok, jobId, type]
      properties:
        ok:
          type: boolean
          const: true
        jobId:
          type: string
        type:
          type: string
          enum: [csv, json, pdf]
        note:
          oneOf:
            - type: string
            - type: 'null'
...